优化抽象语法树节点的继承关系和文件结构
