--- metasequoia_sql/core/node/basic_node.py
@@ -1,63 +1,0 @@
-"""
-抽象语法树（AST）的基础节点类
-"""
-
-import dataclasses
-import enum
-from typing import Optional
-
-from metasequoia_sql.core.node.abc_node import ASTBase
-from metasequoia_sql.core.sql_type import SQLType
-
-__all__ = [
-    "EnumWindowRowType", "ASTWindowRow", "ASTWindowRowExpression",  # 窗口函数的行限制表达式
-]
-
-
-# ---------------------------------------- 窗口函数中的行限制表达式 ----------------------------------------
-
-class EnumWindowRowType(enum.Enum):
-    """窗口函数中的行限制的类型"""
-    PRECEDING = "PRECEDING"
-    CURRENT_ROW = "CURRENT ROW"
-    FOLLOWING = "FOLLOWING"
-
-
-@dataclasses.dataclass(slots=True, frozen=True, eq=True)
-class ASTWindowRow(ASTBase):
-    """窗口函数中的行限制
-
-    unbounded preceding = 当前行的所有行
-    n preceding = 当前行之前的 n 行
-    current row = 当前行
-    n following = 当前行之后的 n行
-    unbounded following = 当前行之后的所有行
-    """
-
-    row_type: EnumWindowRowType = dataclasses.field(kw_only=True)
-    is_unbounded: bool = dataclasses.field(kw_only=True, default=False)
-    row_num: Optional[int] = dataclasses.field(kw_only=True, default=None)
-
-    def source(self, sql_type: SQLType = SQLType.DEFAULT) -> str:
-        """返回语法节点的 SQL 源码"""
-        if self.row_type == EnumWindowRowType.CURRENT_ROW:
-            return "CURRENT ROW"
-        row_type_str = self.row_type.value
-        if self.is_unbounded is True:
-            return f"UNBOUNDED {row_type_str}"
-        return f"{self.row_num} {row_type_str}"
-
-
-@dataclasses.dataclass(slots=True, frozen=True, eq=True)
-class ASTWindowRowExpression(ASTBase):
-    """窗口函数中的行限制表达式
-
-    ROWS BETWEEN {SQLWindowRow} AND {SQLWindowRow}
-    """
-
-    from_row: ASTWindowRow = dataclasses.field(kw_only=True)
-    to_row: ASTWindowRow = dataclasses.field(kw_only=True)
-
-    def source(self, sql_type: SQLType = SQLType.DEFAULT) -> str:
-        """返回语法节点的 SQL 源码"""
-        return f"ROWS BETWEEN {self.from_row.source(sql_type)} AND {self.to_row.source(sql_type)}"
